tugas-heroku/requirements.txt
Pillow==8.4.0
streamlit==1.2.0
tensorflow==2.6.0
keras==2.6.0